+++ server/src/parser.ts
@@ -2,8 +2,12 @@
 
 import {
     Enumeration,
+    JunosSchema,
     Sequence
 } from '../src/junos';
+
+export const prefixPattern = /^[\t ]*(?:set|delete|activate|deactivate)/;
+
 
 export class Node {
     name: string;
@@ -123,7 +127,7 @@ export class Node {
 
     private add_string_node(key: string, raw_children) {
         const [name, description] = this.extract_key(key);
-        const match = name.match(/(\S*)\((.*)\)/);
+        const match: RegExpMatchArray = name.match(/(\S*)\((.*)\)/);
 
         if (!match) {
             this.add(new Node(name, this, raw_children, description));
@@ -213,4 +217,10 @@ export class Parser {
             return node.description;
         }
     }
+}
+
+export function createParser() {
+    const schema = new JunosSchema();
+    const ast = new Node('configuration', null, schema.configuration());
+    return new Parser(ast);
 }